Class1/Class1 (Copy)/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3

# Flask dependencies
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2

# Environment loader (.env)
python-dotenv==1.0.1

# MySQL Database driver (ổn định + nhẹ cho Docker)
pymysql==1.1.0

# Email validation (bạn có sử dụng)
email-validator==2.1.0

# Server chạy Flask khi deploy Docker
gunicorn==21.2.0
